Move test into another package
